mapping and exceptions changes

+++ src/main/java/com/internship/tabulaprocessing/media/service/MediaService.java
@@ -1,7 +1,5 @@
 package com.internship.tabulaprocessing.media.service;
 
-
-
 import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
@@ -15,7 +13,9 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import javax.persistence.EntityNotFoundException;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class MediaService {
@@ -35,31 +35,48 @@ public class MediaService {
         return new ResponseEntity<>(page.toList(), HttpStatus.OK);
     }
 
-    public ResponseEntity<MediaDto> getOne(int id) throws Exception {
-        Media media = mediaRepository.getOne(id);
-        MediaDto mediaDto = mapper.convertValue(media, MediaDto.class);
+    public ResponseEntity<MediaDto> getOne(int id) {
+        Optional<Media> media = mediaRepository.findById(id);
+        if(!media.isPresent()){
+            throw new EntityNotFoundException("Media not found.");
+        }
+        MediaDto mediaDto = mapper.convertValue(media.get(), MediaDto.class);
         return  ResponseEntity.ok(mediaDto);
     }
 
     public ResponseEntity<MediaDto> create(MediaDto mediaDto) {
+
         Media media = mapper.convertValue(mediaDto, Media.class);
         mediaRepository.save(media);
         mediaDto.setId(media.getId());
-        return  new ResponseEntity<MediaDto>(mediaDto, HttpStatus.CREATED);
+
+        return  new ResponseEntity<>(mediaDto, HttpStatus.CREATED);
     }
 
     public ResponseEntity<?> deleteById(int id) {
+
+        Optional<Media> optional = mediaRepository.findById(id);
+
+        if(!optional.isPresent()){
+            throw new EntityNotFoundException(String.format("Media with id of %s was not found!", id));
+        }
+
         mediaRepository.deleteById(id);
-        return ResponseEntity.ok(String.format("Media with id of %s was deleted successfully!",id));
+        return ResponseEntity.ok(String.format("Media with id of %s was deleted successfully!", id));
     }
 
-    public ResponseEntity<MediaDto> update(int id, MediaDto mediaDto) throws Exception {
-        Media media = mediaRepository.getOne(id);
-        if (media == null) throw new Exception("The media you want to update was not found.");
-        media = mapper.convertValue(mediaDto, Media.class);
+    public ResponseEntity<MediaDto> update(int id, MediaDto mediaDto)  {
+        Optional<Media> optional = mediaRepository.findById(id);
+
+        if(!optional.isPresent()){
+            throw new EntityNotFoundException(String.format("Media with id of %s was not found!", id));
+        }
+
+        Media media = mapper.convertValue(mediaDto, Media.class);
         media.setId(id);
         mediaRepository.save(media);
 
         return ResponseEntity.ok(mediaDto);
     }
 }
+
